+++ Changelog.md
@@ -2,6 +2,12 @@
 
 ## Unreleased Changes
 
+ * Add basic HAL types
+
+Add schema vars for the basic HAL types of Resource, Link, and
+Curie. See the [HAL Specification][hal] for full details
+
+[hal]: http://tools.ietf.org/html/draft-kelly-json-hal-06
  * Add support for generating API documentation
 
 To generate project docs using codox, run